refactor: Ditch the Mock folder
