--- src/Uno.UI/Mock/Color.reference.cs
@@ -1,14 +1,0 @@
-﻿using Uno;
-using Uno.Extensions;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using Uno.Disposables;
-using System.Text;
-using System.Threading.Tasks;
-
-
-namespace Microsoft.UI.Xaml
-{
-
-}
