--- src/Uno.UI/Mock/Control.reference.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Microsoft.UI.Xaml.Controls
-{
-	public partial class Control : FrameworkElement
-	{
-
-	}
-}
