--- src/Uno.UI/Mock/Font.wasm.cs
@@ -1,17 +1,0 @@
-﻿using Uno;
-using Uno.Extensions;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using Uno.Disposables;
-using System.Text;
-using System.Threading.Tasks;
-
-
-namespace Microsoft.UI.Xaml
-{
-	public class Font
-	{
-
-	}
-}
