--- src/Uno.UI/Mock/ImageSource.unittests.cs
@@ -1,20 +1,0 @@
-﻿using Uno.Extensions;
-using Uno.Foundation.Logging;
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Text;
-using System.Threading;
-using System.Threading.Tasks;
-using System.Net.Http;
-using System.ComponentModel;
-
-namespace Microsoft.UI.Xaml.Media;
-
-public partial class ImageSource
-{
-	public ImageSource()
-	{
-		InitializeBinder();
-	}
-}
