--- src/Uno.UI/Mock/ImageSource.wasm.cs
@@ -1,9 +1,0 @@
-﻿namespace Microsoft.UI.Xaml.Media;
-
-public partial class ImageSource
-{
-	public ImageSource()
-	{
-		InitializeBinder();
-	}
-}
