--- src/Uno.UI/Mock/ListViewBase.unittests.cs
@@ -1,11 +1,0 @@
-﻿#pragma warning disable 108 // new keyword hiding
-#pragma warning disable 114 // new keyword hiding
-using Microsoft.UI.Xaml.Controls.Primitives;
-
-namespace Microsoft.UI.Xaml.Controls
-{
-	public partial class ListViewBase : Selector
-	{
-
-	}
-}
