--- src/Uno.UI/Mock/ScrollContentPresenter.unittests.cs
@@ -1,21 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Microsoft.UI.Xaml.Controls
-{
-	public partial class ScrollContentPresenter
-	{
-		public void OnMinZoomFactorChanged(float newValue)
-		{
-			throw new NotImplementedException();
-		}
-
-		public void OnMaxZoomFactorChanged(float newValue)
-		{
-			throw new NotImplementedException();
-		}
-	}
-}
