--- src/Uno.UI/Mock/Shape.unittests.cs
@@ -1,7 +1,0 @@
-﻿namespace Microsoft.UI.Xaml.Shapes;
-
-public partial class Shape
-{
-	private void OnFillBrushChanged() { }
-	private void OnStrokeBrushChanged() { }
-}
