--- src/Uno.UI/Mock/StackPanel.reference.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
-
-namespace Microsoft.UI.Xaml.Controls
-{
-	public partial class StackPanel
-	{
-
-	}
-}
