--- src/Uno.UI/Mock/UserControl.reference.cs
@@ -1,8 +1,0 @@
-﻿#pragma warning disable 108 // new keyword hiding
-#pragma warning disable 114 // new keyword hiding
-namespace Microsoft.UI.Xaml.Controls
-{
-	public partial class UserControl : ContentControl
-	{
-	}
-}
