+++ src/Uno.UI/UI/Xaml/Controls/Border/Border.unittests.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Microsoft.UI.Xaml.Controls
+namespace Microsoft.UI.Xaml.Controls;
 {
 	public partial class Border
 	{
