+++ src/Uno.UI/UI/Xaml/Controls/ContentControl/ContentControl.reference.cs
