+++ src/Uno.UI/UI/Xaml/Controls/Panel/Panel.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿
+using System;
 using System.Collections.Generic;
 using System.Text;
 using System.Collections.Specialized;
