+++ src/Uno.UI/UI/Xaml/Controls/Panel/Panel.reference.cs
