+++ src/Uno.UI/UI/Xaml/Media/ImageSource.wasm.cs
@@ -19,6 +19,11 @@ namespace Microsoft.UI.Xaml.Media
 {
 	partial class ImageSource
 	{
+		public ImageSource()
+		{
+			InitializeBinder();
+		}
+
 		partial void InitFromResource(Uri uri)
 		{
 			var path = uri.PathAndQuery.TrimStart("/");
